Use int64 for StreamingReader size (#2688)
It *can* be used on the API for a large streaming upload.  That will be slow
but must still give a correct result.  (Go has much 64-bit confusion, say in
the `len(...)` return type.)

+++ pkg/block/s3/stream_test.go
@@ -108,7 +108,7 @@ func TestS3StreamingReader_Read(t *testing.T) {
 
 			data := &s3a.StreamingReader{
 				Reader:       r,
-				Size:         len(cas.Input),
+				Size:         int64(len(cas.Input)),
 				StreamSigner: v4.NewStreamSigner("us-east-1", s3.ServiceName, sigSeed, keys),
 				Time:         sigTime,
 				ChunkSize:    cas.ChunkSize,
